requirements.txt
asyncpg==0.30.0
redis[hiredis]==5.2.1
pydantic==2.10.3
httpx==0.28.1
structlog==24.4.0
jinja2==3.1.6
weasyprint==68.1
python-dotenv==1.0.1
tenacity==9.0.0
tiktoken==0.8.0
uvicorn[standard]==0.34.0
fastapi==0.125.0
starlette==0.49.1
python-multipart==0.0.26
sse-starlette==2.2.1
python-docx==1.1.2
stripe==11.4.1
python-pptx==1.0.2
openpyxl==3.1.5
defusedxml==0.7.1
